Rename duplicate column names in read_json(orient='split')
