+++ pandas/io/parsers/c_parser_wrapper.py
@@ -30,6 +30,10 @@ from pandas.core.dtypes.dtypes import ExtensionDtype
 
 from pandas.core.indexes.api import ensure_index_from_sequences
 
+from pandas.io.common import (
+    dedup_names,
+    is_potential_multi_index,
+)
 from pandas.io.parsers.base_parser import (
     ParserBase,
     ParserError,
@@ -227,7 +231,10 @@ class CParserWrapper(ParserBase):
         except StopIteration:
             if self._first_chunk:
                 self._first_chunk = False
-                names = self._dedup_names(self.orig_names)
+                names = dedup_names(
+                    self.orig_names,
+                    is_potential_multi_index(self.orig_names, self.index_col),
+                )
                 index, columns, col_dict = self._get_empty_meta(
                     names,
                     self.index_col,
@@ -281,7 +288,9 @@ class CParserWrapper(ParserBase):
             if self.usecols is not None:
                 names = self._filter_usecols(names)
 
-            names = self._dedup_names(names)
+            names = self._dedup_names(
+                names, is_potential_multi_index(names, self.index_col)
+            )
 
             # rename dict keys
             data_tups = sorted(data.items())
@@ -303,7 +312,9 @@ class CParserWrapper(ParserBase):
             # assert for mypy, orig_names is List or None, None would error in list(...)
             assert self.orig_names is not None
             names = list(self.orig_names)
-            names = self._dedup_names(names)
+            names = self._dedup_names(
+                names, is_potential_multi_index(names, self.index_col)
+            )
 
             if self.usecols is not None:
                 names = self._filter_usecols(names)
