+++ pandas/io/parsers/python_parser.py
@@ -37,6 +37,10 @@ from pandas.errors import (
 from pandas.core.dtypes.common import is_integer
 from pandas.core.dtypes.inference import is_dict_like
 
+from pandas.io.common import (
+    dedup_names,
+    is_potential_multi_index,
+)
 from pandas.io.parsers.base_parser import (
     ParserBase,
     parser_defaults,
@@ -259,7 +263,10 @@ class PythonParser(ParserBase):
         columns: Sequence[Hashable] = list(self.orig_names)
         if not len(content):  # pragma: no cover
             # DataFrame with the right metadata, even though it's length 0
-            names = self._dedup_names(self.orig_names)
+            names = dedup_names(
+                self.orig_names,
+                is_potential_multi_index(self.orig_names, self.index_col),
+            )
             # error: Cannot determine type of 'index_col'
             index, columns, col_dict = self._get_empty_meta(
                 names,
@@ -293,7 +300,9 @@ class PythonParser(ParserBase):
         self,
         alldata: list[np.ndarray],
     ) -> tuple[Mapping[Hashable, np.ndarray], Sequence[Hashable]]:
-        names = self._dedup_names(self.orig_names)
+        names = self._dedup_names(
+            self.orig_names, is_potential_multi_index(self.orig_names, self.index_col)
+        )
 
         offset = 0
         if self._implicit_index:
@@ -434,6 +443,7 @@ class PythonParser(ParserBase):
                         if i not in this_unnamed_cols
                     ] + this_unnamed_cols
 
+                    # TODO: Use pandas.io.common.dedup_names instead
                     for i in col_loop_order:
                         col = this_columns[i]
                         old_col = col
